+++ src/ft_fd/ft_putstr_fd.c
@@ -6,7 +6,7 @@
 /*   By: dpoveda- <me@izenynn.com>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/17 17:44:33 by dpoveda-          #+#    #+#             */
-/*   Updated: 2023/03/13 17:19:57 by dpoveda-         ###   ########.fr       */
+/*   Updated: 2023/03/13 19:36:40 by dpoveda-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,5 +18,5 @@ int	ft_putstr_fd(char *s, int fd)
 {
 	if (!s)
 		return (0);
-	return (write(fd, s, ft_strlen(s)));
+	return ((int)write(fd, s, ft_strlen(s)));
 }
