+++ src/ft_printf/ft_printf_str.c
@@ -6,7 +6,7 @@
 /*   By: dpoveda- <me@izenynn.com>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/25 19:00:12 by dpoveda-          #+#    #+#             */
-/*   Updated: 2023/03/13 17:32:24 by dpoveda-         ###   ########.fr       */
+/*   Updated: 2023/03/14 19:45:44 by dpoveda-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,27 +18,27 @@ static size_t	ft_handle_dot(t_print *tab, size_t print_len)
 {
 	tab->dot--;
 	if (print_len > (size_t)tab->dot)
-		print_len = tab->dot;
+		print_len = (size_t)tab->dot;
 	return (print_len);
 }
 
 static void	ft_handle_left_align(t_print *tab, size_t print_len)
 {
-	tab->wd -= print_len;
+	tab->wd -= (int)print_len;
 	while (tab->wd-- > 0)
 	{
 		if (tab->zero)
-			tab->tlen += write(tab->d, "0", 1);
+			tab->tlen += (int)write(tab->d, "0", 1);
 		else
-			tab->tlen += write(tab->d, " ", 1);
+			tab->tlen += (int)write(tab->d, " ", 1);
 	}
 }
 
 static void	ft_handle_right_align(t_print *tab, size_t print_len)
 {
-	tab->wd -= print_len;
+	tab->wd -= (int)print_len;
 	while (tab->wd-- > 0)
-		tab->tlen += write(tab->d, " ", 1);
+		tab->tlen += (int)write(tab->d, " ", 1);
 }
 
 void	ft_printf_str(t_print *tab)
@@ -49,7 +49,7 @@ void	ft_printf_str(t_print *tab)
 	s = va_arg(tab->args, char *);
 	if (!s)
 	{
-		tab->tlen += write(tab->d, "(null)", 6);
+		tab->tlen += (int)write(tab->d, "(null)", 6);
 		return ;
 	}
 	print_len = ft_strlen(s);
@@ -57,7 +57,7 @@ void	ft_printf_str(t_print *tab)
 		print_len = ft_handle_dot(tab, print_len);
 	if ((tab->zero || tab->wd) && !tab->dash)
 		ft_handle_left_align(tab, print_len);
-	tab->tlen += write(tab->d, s, print_len);
+	tab->tlen += (int)write(tab->d, s, print_len);
 	if (tab->wd && tab->dash)
 		ft_handle_right_align(tab, print_len);
 }
